Movement is now complete, robots move and turn normally. havent tested random movement much yet

+++ src/main/java/com/example/robotarenacoursework/Robot.java
@@ -7,12 +7,12 @@ public class Robot {
 
     private String name;
     private int speed;
-    private int XPosition;
-    private int YPosition;
+    private double XPosition;
+    private double YPosition;
     private int sensors;//unsure what data type it should be, perhaps its own class?
     private int direction;
 
-    public Robot(String name, int speed, int XPosition, int YPosition, int sensors, int direction){
+    public Robot(String name, int speed, double XPosition, double YPosition, int sensors, int direction) {
         this.name = name;
         this.speed = speed;
         this.XPosition = XPosition;
@@ -20,34 +20,42 @@ public class Robot {
         this.sensors = sensors;
         this.direction = direction;
     }
+
     //creating a list of getters for other classes to access the robot's data
-    public String getName(){
+    public String getName() {
         return name;
     }
-    public int getSpeed(){
+
+    public int getSpeed() {
         return speed;
     }
-    public int getXPos(){
+
+    public double getXPos() {
         return XPosition;
     }
-    public int getYPos(){
+
+    public double getYPos() {
         return YPosition;
     }
-    public int getDirection(){
+
+    public int getDirection() {
         return direction;
     }
-    public void setDirection(double direction){
+
+    public void setDirection(double direction) {
         this.direction = (int) direction;
     }
-    public int setXPos(int XPosition){
+
+    public double setXPos(double XPosition) {
         return this.XPosition = XPosition;
     }
-    public int setYPos(int YPosition){
+
+    public double setYPos(double YPosition) {
         return this.YPosition = YPosition;
     }
 
     //main method for testing
-    public static void main(String[] args){
+    public static void main(String[] args) {
         Robot robot = new Robot("Robot1", 5, 0, 0, 3, 0);
         System.out.println(robot.getName());
         System.out.println(robot.getSpeed());
